edit: support having a prompt for an edit field
You can create a static text control,
but then you have to align the text in that with the edit field.
You can left align it with extra space showing,
or right align it to ahve it flush with the edit field.
Supporting inline labels negates these issues and allows for readline like setups.